refactor(config): extract the 11 /config routes into src/config_api/routes.py
A pure move, reviewable as the no-op it is: same paths, same handler bodies,
same wire shapes, same status codes. `tests/golden/api_routes.json` changes by
zero bytes — the route inventory flattens included routers, so an extraction
that preserved every path/method/name leaves it byte-identical.

Named `config_api` because `src/config.py` already owns `config`.

Collaborators without a canonical home outside api_server (`get_service`, the
two board-host guards, `_primary_board_entry`, `_reinitialize_board_clients`)
and the two shared accessors other unconverted domains still patch there
(`get_config_manager`, `get_settings_service`) stay as call-time
`from src.api_server import ...` seams. Retiring them is the next commit —
doing it here would have made this move unreviewable.

`src/config_api/routes.py` inherits api_server's lint-ignore set for the same
reason the four earlier extractions did; the conventions pass drops it.

Two assertions in test_status_code_correctness.py scanned `app.routes` for a
path. Top-level `app.routes` does not list a route served by an included
router, so both silently found nothing and errored:

    FAILED tests/test_status_code_correctness.py::
      TestBoardTestUpstreamVerdictsStay200::test_declares_a_response_model
    FAILED tests/test_status_code_correctness.py::
      TestEnableLocalApiUpstreamVerdictsStay200::test_declares_a_response_model

They now look the route up through `build_route_metadata()`, the flattening
accessor, so they keep asserting the same thing after any future extraction.

Full suite: 1 failed, 5918 passed, 2 skipped (the failure is the known
worktree-environmental test_check_image_formats::TestRepositoryIsClean).
Data-dir leak: 0 files.
