+++ src/config_api/__init__.py
@@ -1,0 +1,6 @@
+"""The ``/config`` API domain.
+
+Named ``config_api`` rather than ``config`` because ``src/config.py`` already
+owns that name — it holds the process-wide :class:`~src.config.Config` façade
+these routes read from.
+"""
